requirements.txt
fastapi==0.78.0
sqlmodel==0.0.6
sqlalchemy==1.4.36
sqlalchemy-utils==0.38.2
sqlalchemy-easy-profile==1.2.1
python-decouple==3.6
loguru==0.6.0
arango-orm==0.7.0
cachetools==5.1.0
passlib==1.7.4
bcrypt==3.2.0
Crypto==1.4.1
pycrypto==2.6.1
simplejson==3.17.6
simple_rest_client==1.1.2
jinja2==3.1.2
flask_admin==1.6.0
flask_login==0.6.1
flask_wtf==1.0.1
flask-debugtoolbar==0.13.1
uvicorn==0.17.6
gunicorn==20.1.0
hypercorn[trio]==0.13.2
python-multipart==0.0.5
pymysql==1.0.2
dataset==1.5.2
slugify==0.0.1
brython==3.10.5